+++ src/main/java/at/porscheinformatik/seleniumcomponents/component/HtmlPage.java
@@ -1,6 +1,6 @@
 package at.porscheinformatik.seleniumcomponents.component;
 
-import at.porscheinformatik.seleniumcomponents.AbstractSeleniumPage;
+import at.porscheinformatik.seleniumcomponents.AbstractSeleniumStandalonePage;
 import at.porscheinformatik.seleniumcomponents.SeleniumEnvironment;
 
 /**
@@ -8,7 +8,7 @@ import at.porscheinformatik.seleniumcomponents.SeleniumEnvironment;
  *
  * @author ham
  */
-public class HtmlPage extends AbstractSeleniumPage
+public class HtmlPage extends AbstractSeleniumStandalonePage
 {
 
     private final String url;